Restructured code. Added plot that represent the elevation of the activity's route, with elevation and distance in km. Fixes #29.
