+++ pyopentracks/models/aggregated_stats.py
