+++ pyopentracks/models/auto_import.py
