+++ pyopentracks/models/database.py
