+++ pyopentracks/models/model.py
