+++ pyopentracks/settings.py
