+++ pyopentracks/views/dialogs.py
