+++ pyopentracks/views/graphs.py
@@ -89,3 +89,32 @@ class AggregatedStatsChart:
     def draw_and_show(self):
         self.figure.canvas.draw()
         self.figure.canvas.show()
+
+
+class LinePlot:
+    def __init__(self, xvalues, yvalues):
+        self._xvalues = xvalues
+        self._yvalues = yvalues
+
+        self.figure = Figure()
+        self.figure.subplots()
+
+        self.figure.canvas = FigureCanvas(self.figure)
+
+        self.axes = self.figure.axes[0]
+        self.axes.spines["left"].set_visible(True)
+        self.axes.spines["right"].set_visible(False)
+        self.axes.spines["bottom"].set_visible(True)
+        self.axes.spines["top"].set_visible(False)
+        #self.axes.plot([0, 200],[50, 50],'--g',label='min: '+str(50)+' m')
+        self.axes.fill_between(xvalues, 0, yvalues, facecolor="green", alpha=0.5)
+
+        self.figure.canvas.set_size_request(300, 200)
+        self.figure.canvas.set_has_window(False)
+
+    def get_canvas(self):
+        return self.figure.canvas
+
+    def draw_and_show(self):
+        self.figure.canvas.draw()
+        self.figure.canvas.show()
